Refactor - доработал классы UserStorage & InMemoryUserStorage
