+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,9 +1,11 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.UserAlreadyExistException;
 import ru.yandex.practicum.filmorate.model.User;
+import ru.yandex.practicum.filmorate.service.UserService;
 
 import javax.validation.Valid;
 import java.util.Collection;
@@ -13,65 +15,34 @@ import java.util.HashMap;
 @Slf4j
 @RequestMapping("/users")
 public class UserController {
-    private final HashMap<Integer, User> users = new HashMap<>();
-    private int id = 1;
+    private final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
 
     @GetMapping
     public Collection<User> findAll() {
-        return users.values();
+        return userService.findAll();
     }
 
     @PostMapping
     public User createUser(@Valid @RequestBody User user) {
         log.info("Получен запрос к эндпоинту на добавление пользователя: " + user);
-        findEmailUser(user);
-        createNameUserIsEmpty(user);
-        findMaxIdUsers();
-        user.setId(id);
-        users.put(user.getId(), user);
-        log.info("Пользователь '{}' добавлен", user);
-        id++;
-        return user;
+        return userService.createUser(user);
     }
 
     @PutMapping
-    public User updateUser(@RequestBody User user) throws Exception {
+    public User updateUser(@Valid @RequestBody User user) {
         log.info("Получен запрос к эндпоинту на изменение данных пользователя: '{}'", user);
+        /*
+        //---если валидация по отрицательному id не пойдёт, то нужно раскоментить---//
         if (user.getId() < 0) {
             log.warn("id пользователя {} не может быть меньше 0", user);
             throw new Exception();
         }
-        if (users.containsKey(user.getId())) {
-            users.put(user.getId(), user);
-            log.info("Изменения успешно внесены");
-        } else {
-            createUser(user);
-        }
-        return user;
-    }
-
-    private void findMaxIdUsers() {
-        if (!users.isEmpty()) {
-            for (User userIds : users.values()) {
-                if (userIds.getId() >= id) {
-                    id = userIds.getId() + 1;
-                }
-            }
-        }
-    }
-
-    private void findEmailUser(User user) {
-        for (User o : users.values()) {
-            if (o.getEmail().equals(user.getEmail())) {
-                log.warn("Пользователь с электронной почтой {} уже зарегистрирован.", user.getEmail());
-                throw new UserAlreadyExistException();
-            }
-        }
-    }
-
-    private void createNameUserIsEmpty(User user) {
-        if (user.getName().isBlank()) {
-            user.setName(user.getLogin());
-        }
+        */
+        return userService.updateUser(user);
     }
 }
